refactor(core): from ts to jsdoc
