+++ packages/core/src/plugins/index.d.ts
